Engine - Separate Engine logic in own class and use from Application. Graphics - IndexBuffer support. Assets - WIP AssetImporter (using ALIMER_TOOLS) and plugin for ShaderImporter.

+++ src/Alimer/Graphics/D3D/D3DShaderCompiler.cpp
@@ -108,7 +108,7 @@ namespace alimer
 
         SafeRelease(errorsBlob);
 
-        PODVector<uint8_t> blob(shaderBlob->GetBufferSize());
+        PODVector<uint8_t> blob(static_cast<uint32_t>(shaderBlob->GetBufferSize()));
         memcpy(blob.Data(), shaderBlob->GetBufferPointer(), shaderBlob->GetBufferSize());
         return blob;
     }
